smplx has removed the dtype parameters

+++ FLAME.py
@@ -178,7 +178,7 @@ class FLAME(nn.Module):
         aa_pose = torch.index_select(pose.view(batch_size, -1, 3), 1,
                                      neck_kin_chain)
         rot_mats = batch_rodrigues(
-            aa_pose.view(-1, 3), dtype=dtype).view(batch_size, -1, 3, 3)
+            aa_pose.view(-1, 3)).view(batch_size, -1, 3, 3)
 
         rel_rot_mat = torch.eye(3, device=vertices.device,
                                 dtype=dtype).unsqueeze_(dim=0).expand(batch_size, -1, -1)
@@ -221,7 +221,7 @@ class FLAME(nn.Module):
         vertices, _ = lbs(betas, full_pose, template_vertices,
                                self.shapedirs, self.posedirs,
                                self.J_regressor, self.parents,
-                               self.lbs_weights, dtype=self.dtype)
+                               self.lbs_weights)
 
         lmk_faces_idx = self.lmk_faces_idx.unsqueeze(dim=0).repeat(
             self.batch_size, 1)
